Create the map database:
The binary file sysdata.rda contains two objects: world and
worldHires, which were compiled and extracted from the maps
and mapdata packages, respectively. This is forn internal use only
therefore it is under R/ folder, so it has not to be documented.
The process that generated the data is in the script sysdata.R
inside the (also) newly created data-raw/ folder.