Remove emacs temporary files from Rbuildignore
This was getting me crazy: since I've placed two wildcards to
.Rbuildignore to avoid looking at thosa automatic emacs backup files
(those with #*#) in the file name, I was not managing to run
R CMD build. The command gave me this bizarre error:
[fpmayer@turing GitHub]$ R CMD build FishMaps
* checking for file ‘FishMaps/DESCRIPTION’ ... OK
* preparing ‘FishMaps’:
* checking DESCRIPTION meta-information ... OK
Error in .read_description(ldpath) :
  file 'FishMaps/DESCRIPTION' does not exist
Execution halted
All of this was because of those wildcards. So to avoid all this trouble,
i placed a line in emacs to not generate those files at all, and removed
this things from here.